Feature - screen finish and event sourcing fix

+++ src/main/java/org/master/command/user/handler/UserCommandHandler.java
@@ -8,17 +8,12 @@ import org.master.command.user.commands.UpdatePasswordCommand;
 import org.master.command.user.commands.UpdateUserCommand;
 import org.master.model.user.User;
 import org.master.repository.user.UserRepository;
-import org.master.service.user.UserService;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 @ApplicationScoped
 public class UserCommandHandler {
 
     @Inject
     UserRepository userRepository;
-
-    @Inject
-    ObjectMapper objectMapper; // Inject Jackson's ObjectMapper for JSON conversion
 
     public void handleCreateUserCommand(CreateUserCommand createUserCommand){
 
@@ -35,8 +30,12 @@ public class UserCommandHandler {
     }
 
     public void handleDeleteUserCommand(DeleteUserCommand command) {
-        userRepository.deleteUser(command.id());
-
+        User user = userRepository.findByUuid(command.id());
+        if (user != null) {
+            userRepository.deleteUser(command.id());
+        } else {
+            throw new IllegalArgumentException("User not found with ID: " + command.id());
+        }
     }
 
     public void handleUpdatePasswordCommand(UpdatePasswordCommand command){
